http://bugzilla.zimbra.com/show_bug.cgi?id=9277 bug: 9277
Look for the provisioning setting (account/cos) "zimbraTextAnalyzer" and if it is set, use that analyzer (which must have registered itself with ZimbraAnalyzer: see the AnalyzerSample extension project) for Indexing/Search.

Changing the analyzer for an existing index is likely to cause strange results: make sure you re-index the mailbox unless you know what you're doing.