http://bugzilla.zimbra.com/show_bug.cgi?id=11021 Bug: 11021
Add optional text input field to chooser widget.
It's displayed below the source list. Typing into
it enables the buttons, and hitting Enter in it
adds the content to the target list.

- fix list view reset bug (DwtListView wasn't clearing
selected items during reset) which caused chooser
buttons to be enabled even with empty source list
- add setHandler() method to DwtInputField
- remove empty @return comment in DwtComposite
that was crashing jsdoc.pl