http://bugzilla.zimbra.com/show_bug.cgi?id=2548 bug: 2548
Allow the caller to set the following metadata during a SaveDraft call:

   [l="{folder}"] [f="{flags}"] [t="{tag-list}"] [color="{color}"]

This metadata setting is performed after the draft is saved; if it fails for any reason, the draft will still be saved.  The response contains all the relevant information on the resulting draft, so the caller can figure out the state of the world post-call.

Proxying should now work on SaveDraft.  If an existing draft is specified, we'll proxy to that mailbox.  If it's a new draft, we'll proxy to the specified target folder.  If neither's specified, it's executed in the auth token owner's mailbox.