http://bugzilla.zimbra.com/show_bug.cgi?id=8103 Bug: 8103
Customizable shortcuts.

Via a shortcuts tab in Options, the user
can assign numeric aliases to folders, tags,
and saved searches for use in shortcuts.
The shortcuts tab is disabled for now and
does not appear.

For example, in the file ZmKeys.properties
there is a shortcuts defined as:

    mail.MoveToFolder = M,NNN

Then if the user assigns an alias of 3 to the
folder Test, typing M then 3 will move 
selected messages to Test.

- move tab shortcuts to DwtTabView