http://bugzilla.zimbra.com/show_bug.cgi?id=2712 bug: 2712
When a SearchOperation is executed, we automatically add "-tag:\Deleted" to the end of the query string if the caller is not IMAP.  As a result, searches will not hit IMAP-deleted messages.

When returning SearchConv and GetConv results, omit all messages with the IMAP \Deleted flag set.

The number of non-\Deleted messages in a conversation is returned in the normal "n" attribute on the <c> element.  The actual conversation size is returned in the "total" attribute, which is omitted unless it differs from the "n" attribute value.

Conversation.getMessages() and Conversation.getUnreadMessages() now return a List<Message> instead of a Message[] array.  They still return the full message list, regardless of \Deleted state.

Open issues with \Deleted and the web client:
  - Conversation participant lists still include senders of \Deleted items
  - Item counts on folders (including the Drafts folder) still include \Deleted items