http://bugzilla.zimbra.com/show_bug.cgi?id=9622 bug: 9622
When we create an account, we try to create the mailbox at the same time.  Thus sync will start automatically for new accounts rather than having to wait for the first login and/or server restart.  There's now no need to iterate over all accounts on the system when instantiating the OfflineMailboxManager.