http://bugzilla.zimbra.com/show_bug.cgi?id=11818 bug: 11818
MailItem.setContent() doesn't tell Mailbox.endTransaction() to kill the item's index entry.  That is done automatically because the item's slated for reindexing.

Cleanup of MailItem.getPath().  It now returns null for items that don't have a unique path in the system, e.g. mail messages, contacts, and items under the ROOT (not USER_ROOT, a.k.a. "/") folder.  This is more accurate and doesn't *appear* to break anything...