http://bugzilla.zimbra.com/show_bug.cgi?id=10988 bug: 10988
Significantly simplify the SenderList data structure.  It now contains only the participants and size, so removing a message from the SenderList is no longer possible without regenerating the SenderList.

This makes the code much cleaner and it allows us to get rid of EmailElementCache and friends sooner rather than later.  There may be a small cost in performance, but I really think the old, complicated SenderList was an unnecessary preoptimization.

Also, add an EmailElementCache.EmailType enum to make email XML output cleaner.