http://bugzilla.zimbra.com/show_bug.cgi?id=7425 bug: 7425
Auth tokens are not admin tokens by default; the caller must explicitly request that the authenticated account's admin access be made available for a given auth token.  (admin:Auth does this, while account:Auth does not.)

"Using admin privileges" is an attribute on a Mailbox.OperationContext, usually derived from the auth token of the requestor.  If it's set, we use any admin/domain admin privileges when figuring out whether the caller has sufficient permissions to perform a Mailbox operation.