http://bugzilla.zimbra.com/show_bug.cgi?id=12008 bug: 12008
Sanitize param-value strings used in CN parameter.

Quote the string if it's already quoted or if it contains "," or ":" or ";" characters.

Also quote it if it has non-US-ASCII characters to workaround an Outlook bug even though non-US-ASCII chars don't need quoting according to RFC2445.

Change the double-quote characters to single-quote unless they are used as quote delimiters.  Change all other disallowed characters (CTLs) to question marks.