http://bugzilla.zimbra.com/show_bug.cgi?id=8468 Bug: 8468
Fix timing issue with autocomplete when
GAL autocomplete is turned on, which could
cause input to get corrupted.

This was a problem with the async handling
of the GAL request. Essentially, a lagging GAL
response with stale results was causing the input
to get reset back to a previous state. When a
GAL response comes, we filter the results if more
input has occurred.