http://bugzilla.zimbra.com/show_bug.cgi?id=12364 http://bugzilla.zimbra.com/show_bug.cgi?id=12275 bug: 12364, 12275
Conversation.TagSet should only contain *real* flags and tags.  "Unread" does not qualify, as it's tracked separately.

Recalculate metadata earlier when detecting a migration from the old SenderList to the new SenderList (see bug 10988).  This change may fix bug 12275, but that's a guess because we don't know what triggered bug 12275.