http://bugzilla.zimbra.com/show_bug.cgi?id=12365 bug: 12365
Support the IMAP WITHIN extension.  This merely requires support for the "OLDER" and "YOUNGER" SEARCH terms, both of which take a number of hours as an argument.  We map "OLDER 6" and "YOUNGER 4" to "before:-6h" and "after:-4h" in our search language.

Tighten up nz-number support, so FETCH BODY[0] and FETCH BODY[1]<200.0> will now return BAD because you're not allowed to pass "0" in those locations.  "04" and "00" are also now no longer acceptable numbers (e.g. in "SEARCH LARGER 066").