http://bugzilla.zimbra.com/show_bug.cgi?id=10041 bug: 10041
dumpWindowsTimeZones.pl is a Perl script that exports timezones in Windows registry as either LDIF or iCalendar.  LDIF format is intended to be cut/pasted into ZimbraServer/conf/ldap/zimbra.ldif file.  iCalendar mode exports a valid .ics file that has a bunch of VTIMEZONE objects.  This format is compatible with the timezone registry proposal in IETF, and can potentially replace the TZ list in LDAP.

The script obviously only runs on Windows and does not work under cygwin.  I may just have to upgrade my cygwin version, but we might as well say this script requires a native Win32 perl.  This script is not run during builds.