http://bugzilla.zimbra.com/show_bug.cgi?id=12350 bug: 12350
Added MailboxIndex.addDocument method that takes an array of Lucene Document objects, used when indexing messages with attachments.  This avoids adding the same index redo op to the uncommitted ops list multiple times.

IndexItem.attachToParent remember if it was called already, to avoid attaching itself to the parent multiple times.  This prevents deadlock with another thread calling commit() on the parent op.