http://bugzilla.zimbra.com/show_bug.cgi?id=3726 bug: 3726
Preserve case on IMAP keywords rather than always coercing them to UPPERCASE.  Note that "FOO", "foo", and "Foo" are the same in IMAP-land, so once a Zimbra tag is created, that's the capitalization we'll use for all IMAP folders.

Quoting Mark Crispin:

    Remember, once a keyword is created, all references to that keyword
    are case-independent.

    Put another way, it is impossible to have distinct keywords FOO, foo,
    and FoO simultaneously in the same mailbox.  Those are all 
    representations of the same keyword.

We're permitted to coerce to UPPERCASE, but we're doing this one for the benefit of the Thunderbird folks because we like them.