http://bugzilla.zimbra.com/show_bug.cgi?id=12476 bug: 12476
Switch Mailbox.getModifiedItems() to return item IDs instead of actual MailItems because the number of returned items is bounded only by the size of the mailbox.  Reworked SyncOperation and CollectionSync to use the new interface.

CollectionSync needs to be tweaked to use the batch interface to Mailbox.getItemById() because going 1 at a time is inefficient; SyncOperation is already doing batches of up to 200.