http://bugzilla.zimbra.com/show_bug.cgi?id=12595 bug: 12595
On create/modify of appointment or task, require the organizer field to be set to the mailbox owner, or unset.  Otherwise, throw INVALID_REQUEST exception.

On cancel, throw INVALID_REQUEST if the mailbox owner is not the organizer and recipients are specified.  That is, don't allow non-organizer to send cancellation email to attendees.