http://bugzilla.zimbra.com/show_bug.cgi?id=2249 http://bugzilla.zimbra.com/show_bug.cgi?id=12172 bug: 2249, 12172
The indexing code now takes a parameter "deleteFirst" -- which is set to TRUE when we're re-indexing an existing item, and FALSE when we're not.  This fixes the problem with drafts not being updated in the index.  It also has the side effect of making Contact Creation and Wiki Document Creation much faster (subsequent updates run at the same speed).