http://bugzilla.zimbra.com/show_bug.cgi?id=11295 bug: 11295
When deleting or emptying a folder containing a huge number of unread messages, batch the fetches from the database (batch size is 500) and purge the items from the cache after the "mark read" op is done.  Since the items have already been marked as deleted in the operation's "dirty list", the unread Message items will not be cached there for the duration of the operation either.

Note that this doesn't fix the deeper issue, namely that marking a large folder "read" will still take up lots of memory during the op and thereafter swamp your sessions.  But it's a start.