bug: none
Integration FRANK_to_main of #38740:
Added attribute zimbraSyncWindowSize for possible debugging purpose in the field.  Sometimes
the server returns an item to the client that chokes the client.  From the log everything may
look just fine.  The default sync batch size is 100 for PocketPC, 50 for Nokia, and 7/20 for
Treo650.  It's hard to look for the offending item among 100 items.  When we get into that
situation, this new attribute can allow admin to set the batch size to as small as 1 for the
account.  That way we can find in the log which item actually caused trouble on client.