bug: 12282
Integrate change 38826 from FRANK to main.

Server-side hack to allow web client to see correct meeting start/end times when original values use custom timezone definition web client is unware of.

To enable the hack, do:

# su - zimbra
$ zmlocalconfig -e debug_calendar_force_utc=true
$ tomcat restart

When the hack is enabled, the DTSTART and DTEND value of non-recurring appointments will be converted to UTC date-time.  The hack does not apply to all-day appointments, recurring appointments, or instances of recurring appointments.  It is only good for simple, non-all-day appointments.