bug: 9622, 13186
Package PushChanges into an object just like InitialSync and DeltaSync to make it easier for state to be shared across an entire push session.

Message upload timeout is now proportional to the uploaded message's size.  The rule is 5 seconds plus 1 second per 25K of message size.

Use the "send UID" to recover from connection loss during a SendMsg and thus not re-send a duplicate message on the next change push.