bug: 9622
Resolve generic naming conflicts during change push by using GetItem to determine the conflicting remote item and then <Foo>Action to rename the remote item.  (If this fails, the local item is renamed and an exception is thrown; the push should succeed the next time around.)  Reorganize the code so that all of the PushChanges.syncFoo() commands use the new generic pushRequest() call to update remote items.  You should now be able to swap folder or tag names in the offline client (e.g. rename /foo to /bar and /bar to /foo) and have the changes successfully pushed to the remote server.

Fix OfflineMailbox to record tag creates as syncable actions.  Fix the omission of tags from SyncOperation.elementNameForType() and its reverse.  Don't enforce modify checks locally in the offline client because change numbers jump around due to sync.