http://bugzilla.zimbra.com/show_bug.cgi?id=9028 bug 9028
Allow extra object classes to be specified for a zimbra account.

Eg, to add the samba schema:
 1) cp samba.schema /opt/zimbra/openldap/etc/openldap/schema
 2) edit slapd.conf.in to include this schema file and add any
    indices that samba may require
 3) bounce ldap
 4) zmprov mcf +zimbraAccountExtraObjectClass sambaSamAccount
 5) zmprov ca user6@example.com <password>  sambaSID <sambaSID>

The downside is that sambaSID is a must attribute, so you can no
longer create accounts from the admin console (this is filed as
13672).

Provisioning continues to add amavisAccount by default and there is
no way to turn this off, but everything there is optional.