http://bugzilla.zimbra.com/show_bug.cgi?id=13609 bug: 13609
Integrating change 41198 from FRANK to main

Added <tz> element to <CancelAppointmentRequest> to allow client to provide full definition of a custom time zone referenced in <inst> element when cancelling an instance of a recurring appointment.

CancelAppointmentRequest handler will fix up the <inst> value.  Client may use a timezone that is different from the timezone used in the recurrence series, but for compatibility with various calendar UAs the timezone must be same as the series.  The handler will ensure this.

When setting DTSTART/DTEND in a cancellation invite, use the correct values for the instance.  Previously, they were being set to the first-instance values no matter which instance was being cancelled and this was confusing Outlook.  According to RFC2446 these don't need to be set at all, but omitting them confuses Outlook which ends up displaying the raw MIME in the email reading pane.