http://bugzilla.zimbra.com/show_bug.cgi?id=10120 Integrate 40790 from LAZY to main
-------------------

Bug: 10120

When pushing the current view onto
itself, call _setViewVisible() so that the
elements are sized.

This fixes the problem where a blank
list view occasionally shows up. What
happens is that the height of the app
content element gets set to 0 (I don't
know why since this one is tough to
reproduce).