http://bugzilla.zimbra.com/show_bug.cgi?id=13256 bug: 13256
MailSender.sendMimeMessage() now returns a qualified ItemId any time that a copy was saved to a Sent folder.  At present, this isn't really necessary -- an int would still be fine since the save-to-sent copy is *always* placed in the authenticated user's mailbox.

As a result, in SOAP responses you'll now see an "id" attribute on delegated sends.  Note that the value of that attribute will always be unqualified since the authenticated user always matches the saved item.