http://bugzilla.zimbra.com/show_bug.cgi?id=14424 http://bugzilla.zimbra.com/show_bug.cgi?id=9622 bug: 14424, 9622
Split off Mailbox.delete()'s tombstone collection into its own method.  Override this in the OfflineMailbox so that we can prevent tombstones for being written for locally-created, non-synced tags.

Stop caching the remote password and uri in the OfflineMailbox.  Fetch them from the Account instead.  This makes it so that remote password/uri changes are picked up automatically, without a server restart.