http://bugzilla.zimbra.com/show_bug.cgi?id=14390 Bug: 14390
Removed the code in appt compose that checks
attendee input on any blur event. That code was
fragile and resulted in false positives (where the
"Invalid attendee" dialog popped up).

The good news is that we shouldn't see that
annoying dialog any more. The bad news is that
invalid attendee input (eg a partial address) will
disappear once you change tabs, which may be
confusing.