http://bugzilla.zimbra.com/show_bug.cgi?id=11283 bug: 11283
ParsedDateTime.difference() should special case all-day dates.  The difference between the beginning of the day and the beginning of the next day where the first day is daylight savings time transition date should be computed as "1D" rather than "23H" or "1D1H".

GetAppointmentSummaries response now correctly reports 23 hours for an all-day appointment on the day of transitioning into daylight savings time (March 11, 2007 in US) and 25 hours for the day of transitioning out (November 4, 2007).