http://bugzilla.zimbra.com/show_bug.cgi?id=14620 http://bugzilla.zimbra.com/show_bug.cgi?id=14709 Bug: 14620, 14709
Refactor tree controller handling of change events.
Pulled code that handles colors and checked state
up in ZmTreeController.

Also fixed bug where you couldn't set the color of
an address book or task folder to orange.

Changed constant for ORANGE to 9. A value of 0 or
no value for color gives it the default for the folder
type. Everything that has a color defaults to ORANGE,
except for address books and task folders which default
to GRAY.

If a folder is created with its default color, don't send a 
"color" attribute. If the color is changed to its default,
set it to 0.