http://bugzilla.zimbra.com/show_bug.cgi?id=10325 Bug: 10325
An operation can now have an associated setting
that acts as a precondition. For example, the setting
HTML_COMPOSE_ENABLED is a precondition for the 
operation HTML_FORMAT.

That allows us to pass lists of operations to toolbars
and action menus without first checking all the related
settings. If an operation (eg NEW_CONTACT) depends
on an app and that app is not enabled, the operation will
not be defined and will not appear in the toolbar/action
menu.

Another thing that has been factored into the constructors
of ZmButtonToolBar and ZmActionMenu is figuring out when
it makes sense to show a separator (when there is something
on both sides).