http://bugzilla.zimbra.com/show_bug.cgi?id=9440 bug: 9440
IMAP capabilities can be switched off by setting the multivalued LDAP attrs zimbraImapDisabledCapability and zimbraImapSSLDisabledCapability.  If a capability name is listed in one of these lists, we hide that capability from IMAP connections coming in on the relevant port and disable all the relevant functionality for the connection.

For instance, if you add "zimbraImapDisabledCapability LITERAL+", the string "LITERAL+" will not appear in CAPABILTY responses and attempts to use nonblocking literals (e.g. "{231+}") will fail with a BAD syntax error.