http://bugzilla.zimbra.com/show_bug.cgi?id=14968 bug: 14968
We were using HttpClient without any parameter setting in some places.  The result is SO_TIMEOUT or connection timeout are set to 0 which means to block forever.  Now we set the global default to 6 seconds for both at Offline static initialization.  The two values are mapped to localconfig.xml's http_so_timeout and http_connection_timeout.