http://bugzilla.zimbra.com/show_bug.cgi?id=12508 bug: 12508
Adds an XmlFormatter ("fmt=xml") and JsonFormatter ("fmt=json").  You can now export as XML or JSON from the REST interface.

Extracted the ItemId formatting responsibilities from ZimbraSoapContext and put them in a new ItemIdFormatter class.  Rewrote ToXML to take an ItemIdFormatter (and sometimes an OperationContext) instead of a ZimbraSoapContext.