http://bugzilla.zimbra.com/show_bug.cgi?id=15179 http://bugzilla.zimbra.com/show_bug.cgi?id=15189 bug: 15179, 15189
Force the instantiation of a ParsedDocument object before taking the Mailbox lock when we're creating a Wiki or Document.

Remove the "parent" relationship from Wikis/Documents.  The current callers were all setting it to null, but odd things could happen if someone actually passed in a real MailItem.  We can revisit this if and when we decide we really want to support a full parent/child relationship between wikis and attached documents.