http://bugzilla.zimbra.com/show_bug.cgi?id=10325 Bug: 10325
Part 1 of moving settings and prefs to apps.

Settings IDs are no longer created statically. For now,
they're created by ZmSettings::initialize. In Part 2, many
of them will be created by apps. Same for prefs.

Static references to settings IDs had to be moved to instance
code, generally a singleton constructor.

- instead of creating the singleton settings object lazily,
ZmZimbraMail creates it explicitly and saves it in the app context