http://bugzilla.zimbra.com/show_bug.cgi?id=14535 bug: 14535
On Mac OS X in order to dock something in the main dock the program needs to be a real app.  the JDK on Mac comes with a tool "Jar Bundler" to convert a java class into a mac app resource bundle.  I'm adding a trivial util class so that we can make a mac app to launch offline url.