http://bugzilla.zimbra.com/show_bug.cgi?id=15303 bug: 15303
Unify the NIO and TcpServer IMAP implementations.  ImapHandler is now the abstract base connection handler class and ImapRequest is the base request parser class.  OzImapConnectionHandler and TcpImapHandler inherit off of ImapHandler; OzImapRequest and TcpImapRequest inherit off of ImapRequest.

There's a bit of ick in there because the TcpServer's ProtocolHandler class is abstract (not an interface), so the *base* class has to extend ProtocolHandler for things to work.  If and when it's refactored so that ProtocolHandler is an interface, it'll be TcpImapServer that implements it.  As a result, there's a few cruft functions in OzImapConnectionHandler to make the compiler happy.

Tweaked the NIO version's logging so it directs NIO-related entries to the new zimbra.nio logger.

Moved ImapContinuationException and ImapTerminatedException into TcpImapRequest, as they're TcpServer-specific.  They now inherit off of ImapParseException, so ImapException goes away.

With the unified base class for ImapHandler, some of the hackarounds like the IFindOrCreateTags interface can go away.