http://bugzilla.zimbra.com/show_bug.cgi?id=6937 bug: 6937
Support for <nosession> and <nonotify> in ZMailbox.

If you ask for a folder and the folder hierarchy is not cached -- either because you have <nosession> or <nonotify> active or because there was no <refresh> block (e.g. ZMailbox referencing another user's mailbox, since the delegated notifications bug isn't fixed) -- we go out and fetch tags and folders explicitly.

We also support the case where you have visibility onto only a subset of folders in the remote mailbox via the visible="1" GetFolder option from changelist 47161.

When folders are serialized, we include the content change sequence number in the XML.  We expose this via ZFolder.getContentSequence().