http://bugzilla.zimbra.com/show_bug.cgi?id=10044 bug:10044
support the "is:solo" query operator -- it is useful for finding messages which are not in a conversation. 

Note that this query only looks for 'parent_id=null' -- that will will be true for "solo" messages in the vast majority of cases, however if the user takes an existing conversation and deletes all messages but one, the conversation is not deleted -- therefore those messages will not appear as 'solo' even though they are in conversations w/ only one message.  This is a fairly rare case, and shouldn't affect most people, but it is worth noting.