http://bugzilla.zimbra.com/show_bug.cgi?id=6937 bug: 6937
Move the selected folder and the connection state from the ImapSession to the ImapHandler.  The connection state is now an enum and is calculated dynamically.

When you've selected a folder that's hard-deleted by another session, your connection is unceremoniously dropped.  We were previously unselecting the folder, but that's just going to confuse the client because there's no way of conveying that such a SELECTED -> AUTHENTICATED state change has occurred.