http://bugzilla.zimbra.com/show_bug.cgi?id=14715 bug: 14715
More ways to force the choice of an index for search ("WHERE INDEX_ID IN (...)", "WHERE PARENT_ID = ?").


Due to https://issues.apache.org/jira/browse/DERBY-47, Derby will do a table scan on IN clauses with >1 entry even when there's an index on the column.  To work around that, move the IN clause batch size into the Db class (set to 1 for Derby) and use it almost everywhere we hardcode an IN clause.

Under Derby, this should speed up marking conversations read, moving conversations around, returning conversation results for search, and emptying trash.  There should be no change under MySQL.  (Dynamically generated searches, like the one for "in:inbox" etc., are unaffected.)