http://bugzilla.zimbra.com/show_bug.cgi?id=15437 Bug: 15437
Use a timed action to clear the right-click selection state
of a list view (which resets the row background, and clears
the internal selection list), so that it's done after whatever
action has just been performed.

The bug came up as a result of the change to popping down
a menu as soon as something is selected. While that's the right
way to do things, we used to rely on it not happening until after
the action was done.