http://bugzilla.zimbra.com/show_bug.cgi?id=15645 http://bugzilla.zimbra.com/show_bug.cgi?id=6937 bug: 15645, 6937
Support for IMAP MULTIAPPEND.  Due to our command processing logic, all messages remain in memory until they're added to the target folder.  General cleanup of append and tag creation.

I believe that the NIO IMAP server still does not enforce a maximum *total* length on a request, so this extension does not exacerbate the potential for an OutOfMemoryException from a malicious client.