http://bugzilla.zimbra.com/show_bug.cgi?id=15874 Bug: 15874
Give controls the option to handle dbl click events in isolation.
The problem occurs for controls that handle both single and
double click, when the single click action is heavy. What happens
is that the single click event is processed (needlessly) first, and
then the double click event.

On the first click, we set a 'click pending' flag and start a timer.
If the timer expires before another click arrives, we process the single click.
If a double-click event arrives before the timer expires, then we process the
double-click event.

The downside is that there is a delay before a simple single click event
is processed, since the timer has to expire. Currently, the timer is set
to 300 ms.