http://bugzilla.zimbra.com/show_bug.cgi?id=14630 bug: 14630
The FileUploadServlet now accepts POSTs of raw files.  The POST must specify a "name" on the Content-Type header or a "filename" on the Content-Disposition header or else the upload is ignored.  Files uploaded in this manner receive the same HTTP response body and are accessible in the same way as normal multipart uploads.

Jong did 95% of the work here; I just cannibalized the existing proxied upload fetch code so as not to require in-memory storage of "plain" uploads.