http://bugzilla.zimbra.com/show_bug.cgi?id=15051 Bug: 15051
Be a little more strict with the pattern we use to look
for a valid email address, by requiring something
reasonable at the beginning (nothing, a quote, or a
space). That way an address such as

    a..b..c@xxx.com

won't return a valid addr of "c@xxx.com" (because the
'c' is preceded by a dot).