http://bugzilla.zimbra.com/show_bug.cgi?id=15809 Bug: 15809
Optimize how we handle notifications, mainly mail msg
and conv creates. Targeted for offline client, which can
receive hundreds of conv and msg create notifications
at a time.

The idea is to preprocess the notifications to dismiss all
the ones that won't result in a UI change.

- Changed how notifications are passed to handlers. They
used to be flattened into a list first, but now we just pass
the original hash of notifications.