bug: none
a few weeks ago I did this great big cleanup of the SearchParams parsing -- centralized it in one place, wrote marshalling code so that it could be proxied to a remote server and all that -- and then I forgot to hook it up.  So a lot of search parameters were not being proxied that should have been.  Oops.