http://bugzilla.zimbra.com/show_bug.cgi?id=13431 bug: 13431
Maintain a "total size" count on folders representing the sum of the sizes of all items in the folder.  It's returned in the "s" attribute on the folder.  (Note that mountpoints and search folders will not have this "s" attribute.)

Mailbox version is now 1.3.  This upgrade will force a table scan on each mailbox on the next access after startup.