http://bugzilla.zimbra.com/show_bug.cgi?id=16347 bug: 16347
Added a config key for testing impact of fsync on message blob storage.  By default every blob save/copy is fsynced to disk with a call to FileChannel.force(true) to force file content and metadata to be written to disk before returning.  This is needed for transactional integrity.

To disable the fsync, set debug_disable_message_store_fsync=true in zmlocalconfig.