http://bugzilla.zimbra.com/show_bug.cgi?id=15107 bug: 15107
Add port conflict checking for ModifyServer and ModifyConfig SOAP calls.  It cannot be done by prefModify callback because of race condition when multiple ports are being set/unset in one ModifyServer or ModifyConfig request.

If a server is being modified, check the particular server.
If a global config is being modified, check all the servers detecting if changing ports on the config would result in port conflict on any server.