http://bugzilla.zimbra.com/show_bug.cgi?id=10364 bug: 10364
Support the CONDSTORE exception, with the caveat that SEARCH MODSEQ is disabled until the search query language supports the "modseq:" operator.

Exposed the DB search functionality that let the caller specify that they want to fetch the corresponding ImapMessage with the SearchResult, and expanded it to also include asking for the MOD_METADATA column.  Made sure that when we convert from a MessagePartHit to a MessageHit that this extra cached data goes along.

Fixed a few small IMAP bugs (ID was always returning NO because of parse issues, tags created as a result of a STORE/APPEND weren't applied to the first item).

Mailbox.begingTrackingImap() and Mailbox.begingTrackingSync() no longer take an OperationContext as an argument.