http://bugzilla.zimbra.com/show_bug.cgi?id=16515 bug: 16515
Stopgap solution for BES send-as issue.  For now, a Sender header is generated unless (a) the authenticated sender matches the From header or (b) the sender is an admin using admin privileges.

Longer-term solution is to use real delegated admin, which will detail whether a user can send-as another user.  But this should do for now.