http://bugzilla.zimbra.com/show_bug.cgi?id=16537 bug: 16537
Initial and delta sync of chats, using the new <chat> serialization from change 51898.  Local chat changes are tracked, but there's not yet a command to update/create a chat on the server, so chat changes are not yet pushed back to the server.

Tweak the server version detection code to handle "5.0.0_...".