http://bugzilla.zimbra.com/show_bug.cgi?id=16665 bug: 16665
When we're delta syncing a message or contact that's not present on the offline client, first make sure that it hasn't already been locally hard-deleted.  (In that case, just skip the change, as the subsequent push will take care of the delete.)

When we get a metadata-only change for a contact that's not in the local store, there's something wrong.  To recover, fetch the contact from the remote server and do an initial sync.  (We were just passing off the delta sync notification to initial sync, but for metadata-only changes that doesn't include the contact fields.)