http://bugzilla.zimbra.com/show_bug.cgi?id=15980 bug: 15980
Extended the ProxyServlet so that the caller can specify "&upload=1" and the server will store the fetched document as a FileUpload-type upload rather than returning it inline in the HTTP response.  In upload mode, the response is identical to a FileUpload response, and you can specify "&fmt=raw" on the proxy request to get the "raw" version of a FileUpload response.